+++ python-libraries/data-platform-catalogue/data_platform_catalogue/client/base.py
@@ -8,7 +8,13 @@ from ..entities import (
     DataProductMetadata,
     TableMetadata,
 )
-from ..search_types import ResultType, SearchResponse
+from ..search_types import (
+    MultiSelectFilter,
+    ResultType,
+    SearchFacets,
+    SearchResponse,
+    SortOption,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -65,8 +71,27 @@ class BaseCatalogueClient(ABC):
             ResultType.DATA_PRODUCT,
             ResultType.TABLE,
         ),
+        filters: Sequence[MultiSelectFilter] = (),
+        sort: SortOption | None = None,
     ) -> SearchResponse:
         """
         Wraps the catalogue's search function.
         """
         raise NotImplementedError
+
+    def search_facets(
+        self,
+        query: str = "*",
+        result_types: Sequence[ResultType] = (
+            ResultType.DATA_PRODUCT,
+            ResultType.TABLE,
+        ),
+        filters: Sequence[MultiSelectFilter] = (),
+    ) -> SearchFacets:
+        """
+        Returns facets that can be used to filter the search results.
+        """
+        raise NotImplementedError
+
+    def list_data_products(self) -> SearchResponse:
+        return self.search(count=500, result_types=[ResultType.DATA_PRODUCT])