+++ python-libraries/data-platform-catalogue/data_platform_catalogue/client/datahub/datahub_client.py
@@ -2,13 +2,6 @@ from typing import Sequence
 
 import datahub.emitter.mce_builder as mce_builder
 import datahub.metadata.schema_classes as schema_classes
-from data_platform_catalogue.client.base import (
-    BaseCatalogueClient,
-    CatalogueError,
-    ReferencedEntityMissing,
-    logger,
-)
-from data_platform_catalogue.search_types import ResultType, SearchResponse
 from datahub.emitter.mce_builder import make_data_platform_urn
 from datahub.emitter.mcp import MetadataChangeProposalWrapper
 from datahub.ingestion.graph.client import DatahubClientConfig, DataHubGraph
@@ -31,6 +24,14 @@ from ...entities import (
     DataProductMetadata,
     TableMetadata,
 )
+from ...search_types import (
+    MultiSelectFilter,
+    ResultType,
+    SearchFacets,
+    SearchResponse,
+    SortOption,
+)
+from ..base import BaseCatalogueClient, CatalogueError, ReferencedEntityMissing, logger
 from .search import SearchClient
 
 DATAHUB_DATA_TYPE_MAPPING = {
@@ -325,10 +326,33 @@ class DataHubCatalogueClient(BaseCatalogueClient):
             ResultType.DATA_PRODUCT,
             ResultType.TABLE,
         ),
+        filters: Sequence[MultiSelectFilter] = (),
+        sort: SortOption | None = None,
     ) -> SearchResponse:
         """
         Wraps the catalogue's search function.
         """
         return self.search_client.search(
-            query=query, count=count, page=page, result_types=result_types
+            query=query,
+            count=count,
+            page=page,
+            result_types=result_types,
+            filters=filters,
+            sort=sort,
+        )
+
+    def search_facets(
+        self,
+        query: str = "*",
+        result_types: Sequence[ResultType] = (
+            ResultType.DATA_PRODUCT,
+            ResultType.TABLE,
+        ),
+        filters: Sequence[MultiSelectFilter] = (),
+    ) -> SearchFacets:
+        """
+        Returns facets that can be used to filter the search results.
+        """
+        return self.search_client.search_facets(
+            query=query, result_types=result_types, filters=filters
         )