+++ python-libraries/data-platform-catalogue/data_platform_catalogue/client/datahub/search.py
@@ -4,13 +4,18 @@ from datetime import datetime
 from importlib.resources import files
 from typing import Any, Sequence
 
-from data_platform_catalogue.search_types import (
-    ResultType,
-    SearchResponse,
-    SearchResult,
-)
 from datahub.configuration.common import GraphError
 from datahub.ingestion.graph.client import DataHubGraph
+
+from ...search_types import (
+    FacetOption,
+    MultiSelectFilter,
+    ResultType,
+    SearchFacets,
+    SearchResponse,
+    SearchResult,
+    SortOption,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -23,6 +28,11 @@ class SearchClient:
             .joinpath("search.graphql")
             .read_text()
         )
+        self.facets_query = (
+            files("data_platform_catalogue.client.datahub.graphql")
+            .joinpath("facets.graphql")
+            .read_text()
+        )
 
     def search(
         self,
@@ -33,6 +43,8 @@ class SearchClient:
             ResultType.DATA_PRODUCT,
             ResultType.TABLE,
         ),
+        filters: Sequence[MultiSelectFilter] = (),
+        sort: SortOption | None = None,
     ) -> SearchResponse:
         """
         Wraps the catalogue's search function.
@@ -43,16 +55,27 @@ class SearchClient:
             start = int(page)
 
         types = self._map_result_types(result_types)
+        formatted_filters = self._map_filters(filters)
 
-        variables = {"count": count, "query": query, "start": start, "types": types}
+        variables = {
+            "count": count,
+            "query": query,
+            "start": start,
+            "types": types,
+            "filters": formatted_filters,
+        }
+
+        if sort:
+            variables.update({"sort": sort.format()})
 
         try:
             response = self.graph.execute_graphql(self.search_query, variables)
         except GraphError as e:
-            raise Exception("Unable to execute search") from e
+            raise Exception("Unable to execute search query") from e
 
         page_results = []
         response = response["searchAcrossEntities"]
+        facets = self._parse_facets(response.get("facets", []))
 
         logger.debug(json.dumps(response, indent=2))
 
@@ -71,8 +94,38 @@ class SearchClient:
                 raise ValueError(f"Unexpected entity type: {entity_type}")
 
         return SearchResponse(
-            total_results=response["total"], page_results=page_results
+            total_results=response["total"], page_results=page_results, facets=facets
         )
+
+    def search_facets(
+        self,
+        query: str = "*",
+        result_types: Sequence[ResultType] = (
+            ResultType.DATA_PRODUCT,
+            ResultType.TABLE,
+        ),
+        filters: Sequence[MultiSelectFilter] = (),
+    ) -> SearchFacets:
+        """
+        Returns facets that can be used to filter the search results.
+        """
+        types = self._map_result_types(result_types)
+        formatted_filters = self._map_filters(filters)
+
+        variables = {
+            "query": query,
+            "facets": [],
+            "types": types,
+            "filters": formatted_filters,
+        }
+
+        try:
+            response = self.graph.execute_graphql(self.facets_query, variables)
+        except GraphError as e:
+            raise Exception("Unable to execute facets query") from e
+
+        response = response["aggregateAcrossEntities"]
+        return self._parse_facets(response.get("facets", []))
 
     def _map_result_types(self, result_types: Sequence[ResultType]):
         """
@@ -84,6 +137,14 @@ class SearchClient:
         if ResultType.TABLE in result_types:
             types.append("DATASET")
         return types
+
+    def _map_filters(self, filters: Sequence[MultiSelectFilter]):
+        result = []
+        for filter in filters:
+            result.append(
+                {"field": filter.filter_name, "values": filter.included_values}
+            )
+        return result
 
     def _parse_owner(self, entity: dict[str, Any]):
         """
@@ -141,6 +202,13 @@ class SearchClient:
         tags = self._parse_tags(entity)
         last_updated = self._parse_last_updated(entity)
         name = entity["name"]
+        relationships = entity.get("relationships", {})
+        total_data_products = relationships.get("total", 0)
+        data_products = relationships.get("relationships", [])
+        data_products = [
+            {"id": i["entity"]["urn"], "name": i["entity"]["properties"]["name"]}
+            for i in data_products
+        ]
 
         return SearchResult(
             id=entity["urn"],
@@ -151,6 +219,8 @@ class SearchClient:
             metadata={
                 "owner": owner_name,
                 "owner_email": owner_email,
+                "total_data_products": total_data_products,
+                "data_products": data_products,
             },
             tags=tags,
             last_updated=last_updated,
@@ -176,7 +246,31 @@ class SearchClient:
                 "owner": owner_name,
                 "owner_email": owner_email,
                 "domain": domain,
+                "number_of_assets": properties["numAssets"],
             },
             tags=tags,
             last_updated=last_updated,
         )
+
+    def _parse_facets(self, facets: list[dict[str, Any]]) -> SearchFacets:
+        """
+        Parse the facets and aggregate information from the query results
+        """
+        results = {}
+        for facet in facets:
+            field = facet["field"]
+            if field not in ("domains", "tags", "customProperties", "glossaryTerms"):
+                continue
+
+            options = []
+            for aggregate in facet["aggregations"]:
+                value = aggregate["value"]
+                count = aggregate["count"]
+                entity = aggregate.get("entity") or {}
+                properties = entity.get("properties") or {}
+                label = properties.get("name", value)
+                options.append(FacetOption(value=value, label=label, count=count))
+
+            results[field] = options
+
+        return SearchFacets(results)