+++ terraform/aws/analytical-platform-development/cluster/terraform.tf
@@ -14,7 +14,7 @@ terraform {
     }
     aws = {
       source  = "hashicorp/aws"
-      version = "5.33.0"
+      version = "5.34.0"
     }
     kubernetes = {
       source  = "hashicorp/kubernetes"