+++ containers/jml-extract-lambda/CHANGELOG.md
@@ -9,6 +9,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.3] - 2024-01-31
+
+### Changed
+
+- Bugfix: resolved issue with incorrect formating of log-group
+
+## [1.0.2] - 2024-01-29
+
+### Changed
+
+- Bugfix: resolved issue with incorrect parsing of ID strings
+
 ## [1.0.1] - 2024-01-17
 
 ### Changed